api-visao/requirements.txt
fastapi
uvicorn
python-multipart
requests
transformers==4.40.0
torch
Pillow
pytesseract
deep-translator